move and rename irl.IRLContext to utils.TfEnvContext since I'm going to be using it in more places
